fix: Moving function to utils - cost
